+++ packages/react-ui/components/SidePage/SidePage.tsx
@@ -7,7 +7,7 @@ import { isKeyEscape } from '../../lib/events/keyboard/identifiers';
 import * as LayoutEvents from '../../lib/LayoutEvents';
 import { stopPropagation } from '../../lib/events/stopPropagation';
 import { HideBodyVerticalScroll } from '../../internal/HideBodyVerticalScroll';
-import { ModalStack, ModalStackSubscription } from '../../lib/ModalStack';
+import { ModalStack, ModalStackSubscription } from '../../lib/ModalStack/ModalStack';
 import { RenderContainer } from '../../internal/RenderContainer';
 import { RenderLayer } from '../../internal/RenderLayer';
 import { ZIndex } from '../../internal/ZIndex';
