+++ packages/react-ui/lib/ModalStack/ModalStack.ts
@@ -1,11 +1,12 @@
 import React from 'react';
-import EventEmitter from 'eventemitter3';
 
-import { SidePageProps } from '../components/SidePage';
-import { ModalProps } from '../components/Modal';
+import { SidePageProps } from '../../components/SidePage';
+import { ModalProps } from '../../components/Modal';
+
+import { Emitter } from './Emitter';
 
 export interface StackInfo {
-  emitter: EventEmitter;
+  emitter: Emitter;
   mounted: React.Component[];
 }
 
@@ -17,25 +18,6 @@ export interface ModalStackSubscription {
   remove: () => void;
 }
 
-// there are cases when users use several versions of `@skbkontur/react-ui` in one app.
-// later versions use the `fbemitter` package, which does not have a `removeListener()` method
-// meanwhile, `eventemitter3` does not have a `remove()` method.
-// we didn't take this into account when we replaced `eventemitter3` with `fbemitter`.
-
-// when this problem was found, we released fixes for more popular versions.
-// but not all teams can move to fixed versions, that may cause app crashing.
-// see all details https://github.com/skbkontur/retail-ui/issues/2197
-
-// this code almost eliminates this risk because it adds compatibility with all later versions.
-EventEmitter.prototype.addListener = function (this: { remove(): void } & EventEmitter, ...args) {
-  // declare the `remove()` method as package `fbemitter`
-  this.remove = () => this.removeListener(...args);
-
-  // call an alias of `addListener()`
-  // see https://github.com/primus/eventemitter3/blob/master/index.js#L319
-  return this.on(...args);
-};
-
 export class ModalStack {
   public static add(
     component: React.Component,
@@ -44,14 +26,14 @@ export class ModalStack {
     const { emitter, mounted } = ModalStack.getStackInfo();
     mounted.unshift(component);
     const changeHandler = () => onChange([...mounted]);
-    const _token = emitter.addListener('change', changeHandler);
+    const fallbackFBEmitter = emitter.addListener('change', changeHandler);
     emitter.emit('change');
     return {
       remove: () => {
         // Backward compatible with versions using the `fbemitter` package.
-        if ('remove' in _token) {
+        if ('remove' in fallbackFBEmitter) {
           // @ts-ignore
-          _token.remove();
+          fallbackFBEmitter.remove();
           return;
         }
 
@@ -89,7 +71,7 @@ export class ModalStack {
     return (
       globalWithStack.__ReactUIStackInfo ||
       (globalWithStack.__ReactUIStackInfo = {
-        emitter: new EventEmitter(),
+        emitter: new Emitter(),
         mounted: [],
       })
     );
